Backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
python-multipart
torch
transformers
pillow
datasets
accelerate
torchvision
mlflow
scikit-learn
pandas
